requirements.txt
streamlit
fpdf2
PyMuPDF
python-docx
google-generativeai
spacy-streamlit
python-dateutil

